Add Base Class to extract login, logout, and launch functionality
