+++ src/test/java/com/icanj/directory/IcanjDirectoryTest.java
@@ -1,31 +1,19 @@
 package com.icanj.directory;
 
+import com.icanj.base.BaseClass;
 import org.openqa.selenium.By;
-import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.chrome.ChromeDriver;
-import org.testng.annotations.AfterTest;
-import org.testng.annotations.BeforeTest;
 import org.testng.annotations.Parameters;
 import org.testng.annotations.Test;
 
-public class IcanjDirectoryTest {
-    private WebDriver driver;
+public class IcanjDirectoryTest extends BaseClass {
 
     @Parameters({"user_name", "pass_word"})
-    @BeforeTest(timeOut = 5000)
-    public void launch(String userName, String password) {
-        // Create a new instance of the Chrome driver
-        System.setProperty("webdriver.chrome.driver", "/home/jason/Downloads/chromedriver");
-        driver = new ChromeDriver();
-        driver.get("http://my.icanj.org/");
-        driver.findElement(By.id("j_username")).sendKeys(userName);
-        driver.findElement(By.id("j_password")).sendKeys(password);
-        driver.findElement(By.xpath("//button[contains(text(),'Sign in')]")).click();
-        driver.manage().window().maximize();
-    }
-
-    @Test
-    public void paginationTest() {
+    @Test(priority = 1)
+    public void paginationTest(String userName, String password) {
+        // First, launch the browser
+        launch();
+        // Then, a successful login must be made to continue
+        login(userName, password);
         // Find hyperlink to "Directory" and click on it
         driver.findElement(By.xpath("//a[contains(@href,\"/Directory/Families\")]")).click();
         // Select a page number from the pagination selections at the bottom
@@ -37,7 +25,7 @@ public class IcanjDirectoryTest {
     }
 
     @Parameters({"search_user"})
-    @Test
+    @Test(priority = 2)
     public void searchTest(String searchUser) {
         // Find hyperlink to "Directory" and click on it
         driver.findElement(By.xpath("//a[contains(@href,\"/Directory/Families\")]")).click();
@@ -46,14 +34,10 @@ public class IcanjDirectoryTest {
         // Select the "More Info" button to access full info of the user
         driver.findElement(By.cssSelector("#directory > tbody > tr > td:nth-child(3) > button")).click();
         driver.navigate().back();
+        // Lastly, log out of session
+        logout();
     }
 
-    @AfterTest
-    public void logout() {
-        driver.findElement(By.xpath("//a[contains(@href, \"/j_spring_security_logout\")]")).click();
-        // Close all browser windows to end session
-        driver.quit();
-    }
 }
