+++ src/test/resources/sponsorship.xml
@@ -7,5 +7,5 @@
         <classes>
             <class name="com.icanj.sponsorship.IcanjSponsorshipTest"/>
         </classes>
-    </test> <!-- Test -->
+    </test>
 </suite>
